requirements.txt
blinker==1.7.0
Bootstrap-Flask==2.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==2.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.0.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
python-dotenv==1.0.1
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.0
WTForms==3.0.1
